Bump version, update bstats, drop nms access
