+++ Plugin/pom.xml
@@ -6,7 +6,7 @@
 	<parent>
 		<groupId>com.gmail.filoghost.chestcommands</groupId>
 		<artifactId>chestcommands-parent</artifactId>
-		<version>3.2.0</version>
+		<version>3.3.0-SNAPSHOT</version>
 	</parent>
 
 	<artifactId>chestcommands-plugin</artifactId>
@@ -30,7 +30,7 @@
 
 		<repository>
 			<id>bstats-repo</id>
-			<url>http://repo.bstats.org/content/repositories/releases/</url>
+			<url>https://repo.codemc.org/repository/maven-public/</url>
 		</repository>
 	</repositories>
 
@@ -59,7 +59,7 @@
 		<dependency>
 			<groupId>org.bstats</groupId>
 			<artifactId>bstats-bukkit-lite</artifactId>
-			<version>1.2</version>
+			<version>1.4</version>
 		</dependency>
 	</dependencies>
 
@@ -81,16 +81,9 @@
 			<plugin>
 				<groupId>org.apache.maven.plugins</groupId>
 				<artifactId>maven-shade-plugin</artifactId>
-				<version>3.1.0</version>
+				<version>3.2.1</version>
 				<configuration>
 					<createDependencyReducedPom>false</createDependencyReducedPom>
-					<shadedArtifactAttached>false</shadedArtifactAttached>
-					<artifactSet>
-						<includes>
-							<include>${project.groupId}:*</include>
-							<include>org.bstats</include>
-						</includes>
-					</artifactSet>
 					<filters>
 						<filter>
 							<artifact>*:*</artifact>
